refactor: update prover-scheduler to facilitate testing
