+++ prover-scheduler/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "prover-scheduler"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+tonic = "0.8.1"
+prost = "0.11.0"
+tokio = { version = "1.21.0", features = ["macros", "rt-multi-thread", "signal"] }
+tokio-stream = { version = "0.1" }
+serde_json = "1.0.108"
+
+prover = { path = "../prover" }
+anyhow =  "1.0"
+ethers-providers = { version = "2.0", features = ["ws"] }
+
+serde = "1.0.92"
+serde_derive = "1.0.92"
+uuid = { version = "1.2", features = ["v4", "fast-rng", "macro-diagnostics"] }
+
+#log
+log = "0.4.0"
+env_logger = "0.10"
+
+[build-dependencies]
+tonic-build = "0.8.0"
+
