+++ prover-scheduler/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod scheduler_service;
+
+pub mod service;
