+++ prover-scheduler/src/service/mod.rs
@@ -1,0 +1,2 @@
+pub mod batch_prover_service;
+pub use batch_prover_service::launch_batch_prover_service_with_addr;
