+++ service/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [[bin]] # Bin to run the HelloWorld gRPC server
 name = "batch-prover"
-path = "src/batch_prover.rs"
+path = "src/batch_prover_service.rs"
 
 [dependencies]
 tonic = "0.8.1"
@@ -19,6 +19,7 @@ serde_json = "1.0.108"
 statedb = { path = "../statedb" }
 executor = { path = "../executor" }
 prover = { path = "../prover" }
+prover-scheduler = { path = "../prover-scheduler" }
 plonky = { git = "https://github.com/0xEigenLabs/eigen-zkvm", branch = "main" }
 starky = { git = "https://github.com/0xEigenLabs/eigen-zkvm", branch = "main" }
 algebraic = { git = "https://github.com/0xEigenLabs/eigen-zkvm", branch = "main" }
