+++ service/build.rs
@@ -23,13 +23,13 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             &["proto/src/proto/executor/v1/executor.proto"],
             &["proto/src/proto/executor/v1", "proto/include"],
         )?;
-    tonic_build::configure()
-        .build_server(true)
-        .build_client(true)
-        // .out_dir("proto")
-        .compile(
-            &["proto/src/proto/scheduler/v1/scheduler.proto"],
-            &["proto/src/proto/scheduler/v1", "proto/include"],
-        )?;
+    // tonic_build::configure()
+    //     .build_server(true)
+    //     .build_client(true)
+    //     // .out_dir("proto")
+    //     .compile(
+    //         &["proto/src/proto/scheduler/v1/scheduler.proto"],
+    //         &["proto/src/proto/scheduler/v1", "proto/include"],
+    //     )?;
     Ok(())
 }
